refactor: reshuffle file structure
