+++ lua/grapple/scope.lua
