--- lua/grapple/ui/init.lua
@@ -1,6 +1,0 @@
-local M = {}
-
-M.popup_tags = require("grapple.ui.tags").open
-M.popup_scopes = require("grapple.ui.scope").open
-
-return M
