+++ selene.toml
@@ -1,0 +1,5 @@
+std = "neovim"
+
+[rules]
+global_usage = "allow"
+multiple_statements = "allow"
